capi: renaming and refactoring
